refactor-compact-tiles * Prefix TileRow and DepartureTile
