+++ src/dashboards/Compact/CompactTileRow/CompactTileRow.module.scss
@@ -1,6 +1,6 @@
 @use 'src/assets/styles/Variables';
 
-.Tilerow {
+.CompactTileRow {
     display: flex;
     border-bottom: 2px solid var(--tavla-border-color);
     padding: 1rem 0;
