+++ src/dashboards/Compact/CompactTileRow/PlatformInfo/PlatformInfo.module.scss
