+++ src/dashboards/Compact/CompactTileRow/PlatformInfo/PlatformInfo.tsx
